@@ -19,7 +19,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.19.6 <3.0.0'
+  sdk: '>=2.20.6 <3.0.0'
 
 # Dependencies specify other packages that your package needs in order to work.
 # To automatically upgrade your package dependencies to the latest versions
@@ -36,49 +36,44 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.2
+  intl: ^0.18.0
   camera: ^0.10.5+2
-  intl: ^0.17.0
   flutter_native_splash: ^2.2.19
-  camerawesome: ^1.4.0
-  shared_preferences: ^2.1.2
+  shared_preferences: ^2.2.2
   dropdown_button2: ^2.1.3
-  buddhist_datetime_dateformat_sns: ^1.0.7
   slide_digital_clock: ^1.0.3
   flutter_custom_clippers: ^2.1.0
   calendar_date_picker2: ^0.5.2
-  geolocator: ^9.0.2
-  google_maps_flutter: ^2.3.0
+  geolocator: ^10.1.0
   awesome_dialog: ^3.1.0
-  permission_handler: ^10.3.0
+  permission_handler: ^11.0.1
   url_launcher: ^6.1.11
-  location: ^4.4.0
   flutter_easyloading: ^3.0.5
-  image_picker: ^0.8.9
+  image_picker: ^1.0.4
   flutter_launcher_icons: ^0.13.1
-  app_settings: ^5.0.0
-  any_animated_button: ^0.0.4
-  flutter_bloc: ^8.1.3
-  flutter_local_notifications: ^15.1.0+1
+  flutter_local_notifications: ^16.1.0
   http: ^0.13.6
   flutter_svg: ^2.0.5
   datetime_picker_formfield_new: ^2.1.0
   mailer: ^6.0.1
-  flutter_pin_code_fields: ^2.2.0
-  flutter_pdfview: ^1.3.1
-  dio: ^5.3.2
-  file_picker: ^5.3.1
-  path_provider_android: ^2.1.0
+  dio: ^5.3.3
+  file_picker: ^6.0.0
   path_provider: ^2.1.0
   easy_localization: ^3.0.2
-  get: ^4.6.5
-
-  
-  
-  
+  location: ^5.0.3
+  open_app_settings: ^2.0.1
+  firebase_messaging: ^14.6.9
+  firebase_core: ^2.17.0
+  # flutter_background_service: ^5.0.2
+  # workmanager: ^0.5.2
+  internet_connection_checker_plus: ^1.0.1
+  flutter_screen_lock: ^9.0.1
+  local_auth: ^2.1.7
+  google_maps_flutter: ^2.5.0
 
 dev_dependencies:
 flutter_launcher_icons:
-  android: "launcher_icon"
+  android: "ic_launcher"
   ios: true
   image_path: "assets/Logo_app/hip_logo.png"
   min_sdk_android: 21 # android min sdk min:16, default 21
@@ -130,7 +125,8 @@ flutter:
     - assets/icon_easytime/16w/
     - assets/icon_easytime/32w/
     - assets/icon_easytime/48w/
-    - assets/lang/    
+    - assets/lang/
+    - assets/document/
     # - assets/branding/         
 
   # An image asset can refer to one or more resolution-specific "variants", see
