@@ -1,17 +1,25 @@
-// ignore_for_file: prefer_const_constructors, camel_case_types, prefer_const_constructors_in_immutables, prefer_const_literals_to_create_immutables
+// ignore_for_file: prefer_const_constructors, camel_case_types, prefer_const_constructors_in_immutables, prefer_const_literals_to_create_immutables, unused_import
+import 'dart:developer';
 
+import 'package:awesome_dialog/awesome_dialog.dart';
+import 'package:dio/dio.dart';
 import 'package:easy_localization/easy_localization.dart';
+import 'package:eztime_app/Components/DiaLog/SnackBar/Sanckbar.dart';
 import 'package:eztime_app/Components/TextStyle/StyleText.dart';
+import 'package:eztime_app/Model/Connect_Api.dart';
+import 'package:eztime_app/Model/ResetToken/ResetToken_Model.dart';
 import 'package:eztime_app/Page/Home/Profile/Profile_Page.dart';
 import 'package:eztime_app/Page/Home/Setting/Setting_Page.dart';
 import 'package:eztime_app/Page/Home/Setting/Show_time_information.dart';
 import 'package:eztime_app/Page/Login/Login_Page.dart';
+import 'package:eztime_app/Page/Login/SetDomain_Page.dart';
 import 'package:eztime_app/Page/request/Request_OT_approval.dart';
 import 'package:eztime_app/Page/request/Request_leave.dart';
 import 'package:eztime_app/Page/request/salary_calculation.dart';
 import 'package:eztime_app/Test.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+import 'package:permission_handler/permission_handler.dart%20';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'dart:math' as math;
 
@@ -23,35 +31,97 @@ class MyDrawer extends StatefulWidget {
 }
 
 class _MyDrawerState extends State<MyDrawer> {
-  Future logout(BuildContext context) async {
-    SharedPreferences pref = await SharedPreferences.getInstance();
-    await pref.remove('id');
-    Navigator.of(context).pushAndRemoveUntil(
-        MaterialPageRoute(builder: (context) => Login_Page()),
-        (route) => false);
+  var device_token;
+  var _acessToken;
+  var _resetToken;
+  var _New_Token;
+  var ip;
+  bool _loading = false;
+
+  Future Sharefs() async {
+    SharedPreferences prefs = await SharedPreferences.getInstance();
+    setState(() {
+      _acessToken = prefs.getString('_acessToken');
+      device_token = prefs.getString('_deviceToken');
+      _resetToken = prefs.getString('_resetToken');
+    ip = prefs.getString('ip');
+      log(_resetToken);
+      log('device_token: ${device_token}');
+      log('_acessToken: ${_acessToken}');
+    });
   }
 
-  // on_Goback() {
-  //   setState(() {
+  Future logOutModul() async {
+    try {
+      setState(() {
+        _loading = true;
+      });
+      SharedPreferences prefs = await SharedPreferences.getInstance();
+      String url = '${connect_api().domain}/deleteNotifyToken';
+      var response = await Dio().post(url,
+          data: {'device_token': device_token},
+          options: Options(headers: {
+            "Content-Type": "application/json",
+            "Authorization": "Bearer $_acessToken",
+          }));
+      log('MyDrawerresponse: ${response.statusCode}');
+      if (response.statusCode == 200) {
+        log('MyDrawerresponse: ${response.data}');
+        setState(() {
+          prefs.remove('_acessToken');
+          prefs.remove('_deviceToken');
+          prefs.remove('username');
+          prefs.remove('password');
+          Navigator.of(context).pushAndRemoveUntil(
+              MaterialPageRoute(builder: (context) => Domain_Set_Page()),
+              (route) => false);
+          _loading = false;
+        });
 
-  //   });
-  // }
+        Snack_Bar(
+                snackBarIcon: Icons.check_circle_rounded,
+                snackBarColor: Colors.green,
+                snackBarText: 'homepage.logout')
+            .showSnackBar(context);
+      } else {
+        setState(() {
+          _loading = false;
+          Snack_Bar(
+                  snackBarIcon: Icons.info,
+                  snackBarColor: Colors.red,
+                  snackBarText: 'homepage.logoutfailed')
+              .showSnackBar(context);
+          print('Error: ${response.statusCode}');
+        });
+      }
+    } catch (e) {
+      setState(() {
+        _loading = false;
+        Snack_Bar(
+                snackBarIcon: Icons.info,
+                snackBarColor: Colors.red,
+                snackBarText: 'homepage.catch')
+            .showSnackBar(context);
+      });
+    }
+  }
+  @override
+  void initState() {
+    super.initState();
+    Sharefs();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Drawer(
       backgroundColor: Colors.white,
-      width: 200,
+      width: 250,
       elevation: 8,
+      shape: BeveledRectangleBorder(borderRadius: BorderRadius.all(Radius.zero)),
       child: SingleChildScrollView(
         child: Column(
           children: [
             UserAccountsDrawerHeader(
-              // decoration: BoxDecoration(
-              //   borderRadius: BorderRadius.only(
-              //       bottomLeft: Radius.circular(6),
-              //       bottomRight: Radius.circular(6)),
-              //   color: Colors.blue,
-              // ),
               accountName: Text(
                 "Employee Id.employee Id",
                 style: TextStyle(
@@ -68,6 +138,7 @@ class _MyDrawerState extends State<MyDrawer> {
                   backgroundColor: Colors.white,
                   backgroundImage:
                       AssetImage('assets/background/person-png-icon-29.jpg')),
+                      
             ),
             ListTile(
               minLeadingWidth: 25,
@@ -152,11 +223,35 @@ class _MyDrawerState extends State<MyDrawer> {
             ListTile(
               minLeadingWidth: 25,
               horizontalTitleGap: 0,
-              onTap: () {
-                Navigator.of(context).pop();
-                Navigator.of(context).push(MaterialPageRoute(
-                  builder: (context) => salary_calculation(),
-                ));
+              onTap: () async {
+                await Permission.storage.request();
+                PermissionStatus status = await Permission.storage.status;
+                if (status.isGranted) {
+                  Navigator.of(context).pop();
+                  Navigator.of(context).push(MaterialPageRoute(
+                    builder: (context) => salary_calculation(),
+                  ));
+                } else {
+                  AwesomeDialog(
+                    context: context,
+                    animType: AnimType.scale,
+                    dialogType: DialogType.warning,
+                    title: 'DiaLog.titleinternet'.tr(),
+                    titleTextStyle:
+                        TextStyle(fontWeight: FontWeight.bold, fontSize: 18),
+                    desc: 'DiaLog.Allowaccess'.tr(),
+                    btnCancelText: 'DiaLog.cancle'.tr(),
+                    btnOkText: 'DiaLog.opensetting'.tr(),
+                    btnCancelOnPress: () {
+                      
+                    },
+                    btnOkOnPress: () {
+                      openAppSettings();
+                    },
+                    dismissOnBackKeyPress: false,
+                    dismissOnTouchOutside: false,
+                  )..show();
+                }
               },
               leading: SvgPicture.asset(
                 'assets/icons_Svg/money-bag.svg',
@@ -194,7 +289,8 @@ class _MyDrawerState extends State<MyDrawer> {
               minLeadingWidth: 25,
               horizontalTitleGap: 0,
               onTap: () {
-                logout(context);
+                
+                logOutModul();
               },
               leading: Transform(
                 alignment: Alignment.center,
@@ -216,10 +312,10 @@ class _MyDrawerState extends State<MyDrawer> {
               minLeadingWidth: 25,
               horizontalTitleGap: 0,
               onTap: () {
-                Navigator.of(context).pop();
-                // Navigator.of(context).push(MaterialPageRoute(
-                //   builder: (context) => FlutterPinCodeFields(),
-                // ));
+                // Navigator.of(context).pop();
+                Navigator.of(context).push(MaterialPageRoute(
+                  builder: (context) => test(),
+                ));
               },
               leading: Image.asset(
                 'assets/icon_easytime/1x/icon_X.png',
