@@ -1,5 +1,6 @@
 import 'package:easy_localization/easy_localization.dart';
-import 'package:eztime_app/Components/Buttons/Button.dart';
+import 'package:eztime_app/Components/DiaLog/Buttons/Button.dart';
+import 'package:eztime_app/Components/internet_connection_checker_plus.dart';
 import 'package:flutter/material.dart';
 
 class reset_password extends StatefulWidget {
@@ -10,6 +11,7 @@ class reset_password extends StatefulWidget {
 }
 
 class _reset_passwordState extends State<reset_password> {
+  final _formkey = GlobalKey<FormState>();
   TextEditingController form = TextEditingController();
   TextEditingController form1 = TextEditingController();
   TextEditingController form2 = TextEditingController();
@@ -19,6 +21,7 @@ class _reset_passwordState extends State<reset_password> {
 
   @override
   void initState() {
+    InternetConnectionChecker().checker();
     super.initState();
   }
 
@@ -30,93 +33,168 @@ class _reset_passwordState extends State<reset_password> {
       ),
       floatingActionButton: Buttons(
         title: 'buttons.Save'.tr(),
-        press: () {},
+        press: () {
+          if (_formkey.currentState!.validate()) {
+            
+          }
+        },
       ),
+      key: _formkey,
       floatingActionButtonLocation: FloatingActionButtonLocation.centerFloat,
       body: Center(
-        child: Column(
-          children: [
-            SizedBox(height: 50),
-            Container(
-              width: 250,
-              height: 40,
-              child: TextFormField(
-                controller: form,
-                cursorWidth: 1.5,
-                cursorHeight: 20,
-                obscureText: _obscureText,
-                decoration: InputDecoration(
-                  border: OutlineInputBorder(),
-                  labelText: 'changePass.Password'.tr(),
-                  suffixIcon: IconButton(
-                    icon: Icon(
-                      _obscureText ? Icons.visibility : Icons.visibility_off,
+          child: Column(
+            children: [
+              SizedBox(height: 50),
+              Padding(
+                  padding: const EdgeInsets.symmetric(
+                    horizontal: 40,
+                  ),
+                  child: TextFormField(
+                    obscureText: _obscureText,
+                    controller: form,
+                    decoration: InputDecoration(
+                      contentPadding: EdgeInsets.zero,
+                      enabledBorder: OutlineInputBorder(
+                        borderSide: BorderSide(color: Colors.blue),
+                        borderRadius: BorderRadius.circular(10.0),
+                      ),
+                      border: OutlineInputBorder(
+                        borderSide: BorderSide(color: Colors.blue),
+                        borderRadius: BorderRadius.circular(10.0),
+                      ),
+                      prefixIcon: Icon(
+                        Icons.key_rounded,
+                        color: Colors.blue,
+                      ),
+                      suffixIcon: IconButton(
+                          onPressed: () {
+                            setState(() {
+                              _obscureText = !_obscureText;
+                            });
+                          },
+                          icon: _obscureText
+                              ? Icon(
+                                  Icons.visibility_off_outlined,
+                                  color: Colors.blue,
+                                )
+                              : Icon(
+                                  Icons.visibility,
+                                  color: Colors.blue,
+                                )),
+                      labelText: "รหัสผ่านเดิม",
+                      labelStyle: TextStyle(color: Colors.blue, fontSize: 12),
                     ),
-                    onPressed: () {
-                      setState(() {
-                        _obscureText =
-                            !_obscureText; // เปลี่ยนค่าการซ่อนข้อความ
-                      });
+                    validator: (value) {
+                      if (value!.isEmpty) {
+                        return 'กรุณากรอกรหัสผ่านเดิม!!';
+                      } else {
+                        return null;
+                      }
                     },
                   ),
                 ),
-              ),
-            ),
-            SizedBox(height: 10),
-            Container(
-              width: 250,
-              height: 40,
-              child: TextFormField(
-                controller: form1,
-                cursorWidth: 1.5,
-                cursorHeight: 20,
-                obscureText: _obscureText1,
-                decoration: InputDecoration(
-                  border: OutlineInputBorder(),
-                  labelText: 'changePass.NewPassword'.tr(),
-                  suffixIcon: IconButton(
-                    icon: Icon(
-                      _obscureText1 ? Icons.visibility : Icons.visibility_off,
+                SizedBox(height: 10),
+              Padding(
+                  padding: const EdgeInsets.symmetric(
+                    horizontal: 40,
+                  ),
+                  child: TextFormField(
+                    obscureText: _obscureText1,
+                    controller: form1,
+                    decoration: InputDecoration(
+                      contentPadding: EdgeInsets.zero,
+                      enabledBorder: OutlineInputBorder(
+                        borderSide: BorderSide(color: Colors.blue),
+                        borderRadius: BorderRadius.circular(10.0),
+                      ),
+                      border: OutlineInputBorder(
+                        borderSide: BorderSide(color: Colors.blue),
+                        borderRadius: BorderRadius.circular(10.0),
+                      ),
+                      prefixIcon: Icon(
+                        Icons.key_rounded,
+                        color: Colors.blue,
+                      ),
+                      suffixIcon: IconButton(
+                          onPressed: () {
+                            setState(() {
+                              _obscureText1 = !_obscureText1;
+                            });
+                          },
+                          icon: _obscureText1
+                              ? Icon(
+                                  Icons.visibility_off_outlined,
+                                  color: Colors.blue,
+                                )
+                              : Icon(
+                                  Icons.visibility,
+                                  color: Colors.blue,
+                                )),
+                      labelText: "รหัสผ่านใหม่",
+                      labelStyle: TextStyle(color: Colors.blue, fontSize: 12),
                     ),
-                    onPressed: () {
-                      setState(() {
-                        _obscureText1 =
-                            !_obscureText1; // เปลี่ยนค่าการซ่อนข้อความ
-                      });
+                    validator: (value) {
+                      if (value!.isEmpty) {
+                        return 'กรุณากรอกรหัสผ่านใหม่!!';
+                      } else {
+                        return null;
+                      }
                     },
                   ),
                 ),
-              ),
-            ),
-            SizedBox(height: 10),
-            Container(
-              width: 250,
-              height: 40,
-              child: TextFormField(
-                controller: form2,
-                cursorWidth: 1.5,
-                cursorHeight: 20,
-                obscureText: _obscureText2,
-                decoration: InputDecoration(
-                  border: OutlineInputBorder(),
-                  labelText: 'changePass.ConfirmPassword'.tr(),
-                  suffixIcon: IconButton(
-                    icon: Icon(
-                      _obscureText2 ? Icons.visibility : Icons.visibility_off,
+                SizedBox(height: 10),
+              Padding(
+                  padding: const EdgeInsets.symmetric(
+                    horizontal: 40,
+                  ),
+                  child: TextFormField(
+                    obscureText: _obscureText2,
+                    controller: form2,
+                    decoration: InputDecoration(
+                      contentPadding: EdgeInsets.zero,
+                      enabledBorder: OutlineInputBorder(
+                        borderSide: BorderSide(color: Colors.blue),
+                        borderRadius: BorderRadius.circular(10.0),
+                      ),
+                      border: OutlineInputBorder(
+                        borderSide: BorderSide(color: Colors.blue),
+                        borderRadius: BorderRadius.circular(10.0),
+                      ),
+                      prefixIcon: Icon(
+                        Icons.key_rounded,
+                        color: Colors.blue,
+                      ),
+                      suffixIcon: IconButton(
+                          onPressed: () {
+                            setState(() {
+                              _obscureText2 = !_obscureText2;
+                            });
+                          },
+                          icon: _obscureText2
+                              ? Icon(
+                                  Icons.visibility_off_outlined,
+                                  color: Colors.blue,
+                                )
+                              : Icon(
+                                  Icons.visibility,
+                                  color: Colors.blue,
+                                )),
+                      labelText: "ยืนยันรหัสผ่านใหม่",
+                      labelStyle: TextStyle(color: Colors.blue, fontSize: 12),
                     ),
-                    onPressed: () {
-                      setState(() {
-                        _obscureText2 =
-                            !_obscureText2; // เปลี่ยนค่าการซ่อนข้อความ
-                      });
+                    validator: (value) {
+                      if (value!.isEmpty) {
+                        return 'กรุณากรอกรหัสผ่านใหม่!!';
+                      } else {
+                        return null;
+                      }
                     },
                   ),
                 ),
-              ),
-            ),
-          ],
+            ],
+          ),
         ),
-      ),
-    );
+      );
+    
   }
 }
